src/com/amazon/CoinChange.java
package com.amazon;

import java.util.Scanner;

/**
 * Created by kaibohao on 2016-11-16.
 */
public class CoinChange {

    public long getChangeWays(int total, int coins[]) {
        long temp[][] = new long[coins.length + 1][total + 1];
        for (int i = 0; i <= coins.length; i++) {
            temp[i][0] = 1;
        }
        for (int i = 1; i <= coins.length; i++) {
            for (int j = 1; j <= total; j++) {
                if (coins[i - 1] > j) {
                    temp[i][j] = temp[i - 1][j];
                } else {
                    temp[i][j] = temp[i][j - coins[i - 1]] + temp[i - 1][j];
                }
            }
        }
        return temp[coins.length][total];
    }

    public static void main(String[] args) {
        Scanner in = new Scanner(System.in);
        int n = in.nextInt();
        int m = in.nextInt();
        int coins[] = new int[m];
        for (int coins_i = 0; coins_i < m; coins_i++) {
            coins[coins_i] = in.nextInt();
        }
        CoinChange coinChange = new CoinChange();
        long changes = coinChange.getChangeWays(n, coins);
        System.out.println(changes);
    }
}
